LPD-43135 Inline JspCompiler into CompilerWrapper
